Add check before droping drug_id constraints

+++ app/database/migrations/2017_09_03_163018_remove_drug_id_column.php
@@ -12,11 +12,13 @@ class RemoveDrugIdColumn extends Migration {
 	 */
 	public function up()
 	{
-		Schema::table('disc_diffusion_guidelines', function(Blueprint $table)
-		{
-			$table->dropForeign('disc_diffusion_guidelines_drug_id_foreign');
-			$table->dropColumn('drug_id');
-		});
+		if(Schema::hasColumn('disc_diffusion_guidelines', 'drug_id'))  {
+			Schema::table('disc_diffusion_guidelines', function(Blueprint $table)
+			{
+				$table->dropForeign('disc_diffusion_guidelines_drug_id_foreign');
+				$table->dropColumn('drug_id');
+			});
+		}
 	}
 
 	/**
